Extract packages: text_extractor, assist_structure
